LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/stroke-linejoin-expected.txt


PASS Can set 'stroke-linejoin' to CSS-wide keywords
FAIL Can set 'stroke-linejoin' to var() references assert_equals: expected 2 but got 1
FAIL Can set 'stroke-linejoin' to the 'crop' keyword Invalid values
FAIL Can set 'stroke-linejoin' to the 'arcs' keyword Invalid values
PASS Can set 'stroke-linejoin' to the 'miter' keyword
PASS Can set 'stroke-linejoin' to the 'bevel' keyword
PASS Can set 'stroke-linejoin' to the 'round' keyword
FAIL Can set 'stroke-linejoin' to the 'stupid' keyword Invalid values
PASS Setting 'stroke-linejoin' to a length throws TypeError
PASS Setting 'stroke-linejoin' to a percent throws TypeError
PASS Setting 'stroke-linejoin' to a time throws TypeError
PASS Setting 'stroke-linejoin' to an angle throws TypeError
PASS Setting 'stroke-linejoin' to a flexible length throws TypeError
PASS Setting 'stroke-linejoin' to a number throws TypeError
PASS Setting 'stroke-linejoin' to a URL throws TypeError
PASS Setting 'stroke-linejoin' to a transform throws TypeError
FAIL 'stroke-linejoin' does not supported 'crop bevel' assert_not_equals: Unsupported value must not be null got disallowed value null
FAIL 'stroke-linejoin' does not supported 'round arcs' assert_not_equals: Unsupported value must not be null got disallowed value null
